Move fetch planning logic into Consumers
So far the fetch rewrite logic was part of the SelectStatementPlanner.
This meant that fetch was only possible as a last step.

By moving the logic from the SelectStatementPlanner into the respective
consumers it is possible to have intermediate fetches.
With this commit this is already the case for non-grouping global
aggregations on subqueries.

In the future this will also allow efficient non-rewritable simple
subqueries.